backend/requirements.txt
asgiref==3.8.1 #django
boto3==1.35.74
botocore==1.35.74 #boto3
click==8.1.7 #uvicorn 
dj-database-url==2.3.0
Django==5.1.3 
django-cors-headers==4.6.0 
django-environ==0.11.2
django-rename-app==0.1.7
django-storages==1.14.4
djangorestframework==3.15.2 #djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1 
gunicorn==23.0.0
h11==0.14.0 #uvicorn
jmespath==1.0.1 #boto3
packaging==24.2 #gunicorn
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1 #djangorestframework-simplejwt
python-dateutil==2.9.0.post0 #boto3
# pytz==2024.2
s3transfer==0.10.4
six==1.16.0 #boto3
sqlparse==0.5.2 #django
typing_extensions==4.12.2 #dj-database-url
urllib3==2.2.3 #boto3
uvicorn==0.32.1
whitenoise==6.8.2
